Full refactor GS+
